requirements.txt
dgl==0.8.0
einops==0.3.0
imbalanced_learn==0.8.0
imblearn==0.0
matplotlib==3.5.1
numpy==1.21.0
PyYAML==6.0
scikit_image==0.18.1
scikit_learn==1.1.1
scipy==1.6.3
SimpleITK==2.1.1.2
skimage==0.0
swin_transformer_pytorch==0.4.1
timm==0.4.9
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.1
vit_pytorch==0.22.0
